HEEDLS-553 Rename old DateValidator to OldDateValidator
Keep this helper class around for existing usage in Learning Portal (to be refactored later in 560)
