+++ DigitalLearningSolutions.Web/Controllers/LearningPortalController/Current.cs
@@ -56,7 +56,7 @@
                 return RedirectToAction("Current");
             }
 
-            var validationResult = DateValidator.ValidateDate(day, month, year);
+            var validationResult = OldDateValidator.ValidateDate(day, month, year);
             if (!validationResult.DateValid)
             {
                 return RedirectToAction("SetCurrentCourseCompleteByDate", new { id, day, month, year });
@@ -92,7 +92,7 @@
 
             if (day != null && month != null && year != null)
             {
-                model.CompleteByValidationResult = DateValidator.ValidateDate(day.Value, month.Value, year.Value);
+                model.CompleteByValidationResult = OldDateValidator.ValidateDate(day.Value, month.Value, year.Value);
             }
 
             return View("Current/SetCompleteByDate", model);
