+++ DigitalLearningSolutions.Web/Controllers/LearningPortalController/SelfAssessment.cs
@@ -140,7 +140,7 @@
                 return RedirectToAction("Current");
             }
 
-            var validationResult = DateValidator.ValidateDate(day, month, year);
+            var validationResult = OldDateValidator.ValidateDate(day, month, year);
             if (!validationResult.DateValid)
             {
                 return RedirectToAction("SetSelfAssessmentCompleteByDate", new { selfAssessmentId, day, month, year });
@@ -165,7 +165,7 @@
 
             if (day != null && month != null && year != null)
             {
-                model.CompleteByValidationResult = DateValidator.ValidateDate(day.Value, month.Value, year.Value);
+                model.CompleteByValidationResult = OldDateValidator.ValidateDate(day.Value, month.Value, year.Value);
             }
 
             return View("Current/SetCompleteByDate", model);
